refactor: decouple JPA Entities from Domain Entities
Since this is a point that caught my interest, I decided to decouple the JPA Entities.
This is something I was always thinking about but didn't want to introduce an antipattern.
This rewrite is the chance to try this pattern out.

After reading through some discussions, it is totally discusable and a valid/clean aprroach,
see the following links:
https://twitter.com/bitboss/status/1005518596276187137?lang=de
https://stackoverflow.com/questions/46227697/should-jpa-entities-and-ddd-entities-be-the-same-classes
https://stackoverflow.com/questions/57148900/jpa-entity-responsibility-in-domain-driven-design
https://stackoverflow.com/questions/31400432/ddd-domain-entities-vo-and-jpa