refactor: migrate to a single-module project
Turned out that the layer-archtiecture brought too much complexity, I decided to go back to a simple single-module archtiecture, which shouldn't be a problem as we're using microservices and won't have an exhausting codebase. We could refactor that at a later point on, but imo the experiment doesn't make much sense now.